### 2024-03-21 #6510 @alexlancaster
@Kevin-Mattheus-Moerman : `CONTRIBUTING.md` added in https://github.com/amberbiology/scribl/commit/ab67760411e998ac6c405a16543dd7d8d6f8a6a9
