### 2024-03-27 #6510 @alexlancaster
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> MISSING DOIs
>
> - No DOI given, and none found for title: The Kappa Language and Kappa Tools
> ```

There is no DOI for this software, the official documentation: https://kappalanguage.org/documentation recommends citing using the following, which we did:

`P. Boutillier, J. Feret, J. Krivine, and W. Fontana. The Kappa Language and Tools (version of <insert date on cover>), kappalanguage.org.`

> ```
> - No DOI given, and none found for title: Exploring network structure, dynamics, and functio...
> ```

It looks like this is part of a DOE conference series, and the official URL: https://www.osti.gov/biblio/960616 provides BibTeX reference, but does not provide a DOI.
