### 2024-04-05 #6510 @atrisovic
Hi @benlansdell,
I am writing to invite you to review a paper titled "scribl: A system for the semantic capture of relationships in biological literature", which is currently in the review process at the Journal of Open Source Software (JOSS). Your insights and feedback on this work would be valuable and much appreciated. To help with the review, we have [a reviewers' checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#review-checklist) with useful guidelines.
Thank you!
Ana Trisovic, JOSS Topic Editor
