### 2024-04-09 aa2d6e4
Update matplotlib requirement from <=3.8.3 to <=3.8.4 (#10)
Updates the requirements on [matplotlib](https://github.com/matplotlib/matplotlib) to permit the latest version.
- [Release notes](https://github.com/matplotlib/matplotlib/releases)
- [Commits](https://github.com/matplotlib/matplotlib/compare/v0.91.2...v3.8.4)

---
updated-dependencies:
- dependency-name: matplotlib
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
