+++ pyproject.toml
@@ -24,7 +24,7 @@ dependencies = [
     "pandas <= 2.2.1",
     "pyzotero <= 1.5.18",
     "networkx <= 3.2.1",
-    "matplotlib <= 3.8.3"
+    "matplotlib <= 3.8.4"
 ]
 
 [project.optional-dependencies]
