+++ pyproject.toml
@@ -23,7 +23,7 @@ dependencies = [
     "pyparsing <= 3.1.2",
     "pandas <= 2.2.1",
     "pyzotero <= 1.5.18",
-    "networkx <= 3.2.1",
+    "networkx <= 3.3",
     "matplotlib <= 3.8.4"
 ]
 
