### 2024-04-19 #6645 @editorialbot
[REVIEW]: scribl: A system for the semantic capture of relationships in biological literature
**Submitting author:** @alexlancaster (<a href="http://orcid.org/0000-0002-0002-9263">Alexander Lancaster</a>)
**Repository:** https://github.com/amberbiology/scribl
**Branch with paper.md** (empty if default branch): manuscript
**Version:** v0.8.0
