### 2024-05-20 7cecdda
--- updated-dependencies: - dependency-name: matplotlib   dependency-type: direct:production ...
Signed-off-by: dependabot[bot] <support@github.com>
