+++ pyproject.toml
@@ -24,7 +24,7 @@ dependencies = [
     "pandas <= 2.2.2",
     "pyzotero <= 1.5.18",
     "networkx <= 3.3",
-    "matplotlib <= 3.8.4"
+    "matplotlib <= 3.9.0"
 ]
 
 [project.optional-dependencies]
