### 2024-05-22 5209309
--- (#13)
updated-dependencies:
- dependency-name: matplotlib
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Alex Lancaster <alexlancaster@users.noreply.github.com>
