### 2024-06-24 #6645 @mhucka
I have some comments for the authors, but I'm not sure what the protocol is for doing that. (I did read https://joss.readthedocs.io/ but didn't find instructions regarding this – perhaps I missed them.) So, I'm writing them here.

## Summary

The paper describes a novel and interesting approach to creating a semantic network of relationships between entities, processes, and other concepts in literature about systems biology. The approach leverages a commonly used bibliographic database (Zotero) and enables groups of researchers to curate information and collaboratively build up the network of relationships. The paper reads well; the documentation accompanying the software (scribl) is extensive and detailed. I think the paper and software tick all the boxes for JOSS, and constitute a new and fresh idea.

## Typos & such

- p. 1, line 16: I think the word "for" needs to be inserted after "suitable".
- p. 3, line 58: "celluar" → "cellular".

## Suggestions (optional)

- In the section on related work ("What scribl is not"), I think it would enhance the paper to add a sentence or two contrasting scribl to Cytoscape. Although Cytoscape is indirectly referenced in the abstract by referencing the paper by Suderman & Hallett, I suspect that many readers working in this domain are likely to know about Cytoscape for graph visualization, and may wonder how scribl relates to it. IMHO, a direct explanation in the authors' own words would be preferable to making readers work it out for themselves.
- Although the paper shows examples of scribl syntax used in adding tags to Zotero entries, it's not immediately obvious to readers how to make use of this data. The README for the repository does include examples of how to actually run scribl; still, in reading this paper, I kept thinking "okay, but what do I _do_ with this?" and had to go to the repository to find out. So I think another enhancement to the paper would be to add just one example of running scribl on the command line. It can be simply words to the effect of "After installing and configuring scribl, it can be simply invoked as a shell command, as shown in the following basic example", followed by a code block with a one-line scribl shell command. It would make the paper slightly more self-contained and help communicate to readers that usage is simple.
- In figure 2, the font size of the text inside the boxes of the figure is _very_ small, verging on microscopically small. Yes, readers can zoom in to read it, but it looks like there's enough room to enlarge the text a little bit. It doesn't have to be made huge, but IMHO a small enlargement would help. Figure 1 could use the same treatment; figure 3 is okay.
