### 2024-06-27 9af24e9
fix typos noted in @mhucka's review
https://github.com/openjournals/joss-reviews/issues/6645
