### 2024-06-27 3be6ba6
update github action version to v4 (v1 deprecated)
