### 2024-06-27 668ea81
joss review: https://github.com/openjournals/joss-reviews/issues/6645
implement item 2 in suggestion for sample command-line
