### 2024-06-27 109ade3
trim the example command-line option so it fits
