### 2024-06-27 293fc4a
adjust figure sizes
