+++ paper/paper.md
@@ -51,7 +51,7 @@ that they describe. Examples of relationships for each of the five
 basic entities for a single article are shown in
 \autoref{fig:scribl-schema}.
 
-![The scribl schema comprises a hierarchy of five basic entities: `article`, `category`, `resource`, `process`, and `agent`. Here we depict an example network of entities and possible relationships for a single `article`.\label{fig:scribl-schema}](scribl-schema.png){ width=80% }
+![The scribl schema comprises a hierarchy of five basic entities: `article`, `category`, `resource`, `process`, and `agent`. Here we depict an example network of entities and possible relationships for a single `article`.\label{fig:scribl-schema}](scribl-schema.png){ width=85% }
 
 A curator can add `scribl` statements as tags to each article
 in a literature database, to represent aspects of the causal relationships
@@ -85,7 +85,7 @@ use in graph database platforms. `scribl` also supports the
 incremental updating of the graph database as new Zotero entries come
 in (\autoref{fig:scribl-workflow}).
 
-![Two major workflows for the `scribl` software: creating a new graph database (left) and updating an existing one (right). The workflow contains a step that identifies possible syntactic errors in `scribl` statements so that they can be fixed in the Zotero database before database generation. Note that "Zotero csv export" could be replaced by a query to a remote Zotero library  \label{fig:scribl-workflow}](scribl-workflow.png)
+![Two major workflows for the `scribl` software: creating a new graph database (left) and updating an existing one (right). The workflow contains a step that identifies possible syntactic errors in `scribl` statements so that they can be fixed in the Zotero database before database generation. Note that "Zotero csv export" could be replaced by a query to a remote Zotero library  \label{fig:scribl-workflow}](scribl-workflow.png){ width=105% }
 
 `scribl` functions can be accessed programatically through writing a
 Python script that calls the `scribl` API, or via the included
