### 2024-06-27 c2a8624
use `convert -trim` to remove all whitespace
