### 2024-06-27 a0a5747
increase Fig 1 size, move text to accomodate
