+++ paper/paper.md
@@ -45,13 +45,13 @@ graph-based relationship queries.
 
 # The `scribl` language
 
+![The scribl schema comprises a hierarchy of five basic entities: `article`, `category`, `resource`, `process`, and `agent`. Here we depict an example network of entities and possible relationships for a single `article`.\label{fig:scribl-schema}](scribl-schema.png){ width=90% }
+
 The language was designed for the curation of scientific articles to
 document the relationships between biological agents and processes
 that they describe. Examples of relationships for each of the five
 basic entities for a single article are shown in
 \autoref{fig:scribl-schema}.
-
-![The scribl schema comprises a hierarchy of five basic entities: `article`, `category`, `resource`, `process`, and `agent`. Here we depict an example network of entities and possible relationships for a single `article`.\label{fig:scribl-schema}](scribl-schema.png){ width=82% }
 
 A curator can add `scribl` statements as tags to each article
 in a literature database, to represent aspects of the causal relationships
