### 2024-06-27 425f55e
bump figure sizes a little more
