### 2024-06-27 052903f
add context for Cytoscape as suggestion in review
https://github.com/openjournals/joss-reviews/issues/6645
