+++ paper/paper.bib
@@ -133,6 +133,27 @@
        year = {2015}
 }
 
+
+@article{shannon_cytoscape_2003,
+	abstract = {Cytoscape is an open source software project for integrating biomolecular interaction networks with high-throughput expression data and other molecular states into a unified conceptual framework. Although applicable to any system of molecular components and interactions, Cytoscape is most powerful when used in conjunction with large databases of protein-protein, protein-DNA, and genetic interactions that are increasingly available for humans and model organisms. Cytoscape's software Core provides basic functionality to layout and query the network; to visually integrate the network with expression profiles, phenotypes, and other molecular states; and to link the network to databases of functional annotations. The Core is extensible through a straightforward plug-in architecture, allowing rapid development of additional computational analyses and features. Several case studies of Cytoscape plug-ins are surveyed, including a search for interaction pathways correlating with changes in gene expression, a study of protein complexes involved in cellular recovery to DNA damage, inference of a combined physical/functional interaction network for Halobacterium, and an interface to detailed stochastic/kinetic gene regulatory models.},
+	author = {Shannon, Paul and Markiel, Andrew and Ozier, Owen and Baliga, Nitin S. and Wang, Jonathan T. and Ramage, Daniel and Amin, Nada and Schwikowski, Benno and Ideker, Trey},
+	doi = {10.1101/gr.1239303},
+	issn = {1088-9051, 1549-5469},
+	journal = {Genome Research},
+	language = {en},
+	month = jan,
+	number = {11},
+	pages = {2498--2504}
+	pmid = {14597658},
+	shorttitle = {Cytoscape},
+	title = {Cytoscape: {A} {Software} {Environment} for {Integrated} {Models} of {Biomolecular} {Interaction} {Networks}},
+	url = {https://genome.cshlp.org/content/13/11/2498},
+	urldate = {2024-06-27},
+	volume = {13},
+	year = {2003}
+}
+
+
 @article{suderman_tools_2007,
        abstract = {Many tools exist for visually exploring biological networks including well-known examples such as Cytoscape, VisANT, Pathway Studio and Patika. These systems play a key role in the development of integrative biology, systems biology and integrative bioinformatics. The trend in the development of these tools is to go beyond ‘static’ representations of cellular state, towards a more dynamic model of cellular processes through the incorporation of gene expression data, subcellular localization information and time-dependent behavior. We provide a comprehensive review of the relative advantages and disadvantages of existing systems with two goals in mind: to aid researchers in efficiently identifying the appropriate existing tools for data visualization; to describe the necessary and realistic goals for the next generation of visualization tools. In view of the first goal, we provide in the Supplementary Material a systematic comparison of more than 35 existing tools in terms of over 25 different features.Contact:  msuder@mcb.mcgill.caSupplementary information: Supplementary data are available at Bioinformatics online.},
        author = {Suderman, Matthew and Hallett, Michael},
