### 2024-06-27 7935e04
fix bibtex syntax error
https://github.com/openjournals/joss-reviews/issues/6645
