+++ paper/paper.bib
@@ -143,7 +143,7 @@
 	language = {en},
 	month = jan,
 	number = {11},
-	pages = {2498--2504}
+	pages = {2498--2504},
 	pmid = {14597658},
 	shorttitle = {Cytoscape},
 	title = {Cytoscape: {A} {Software} {Environment} for {Integrated} {Models} of {Biomolecular} {Interaction} {Networks}},
