### 2024-06-27 #6645 @alexlancaster
Thanks for the review, @mhucka!  All responses are below and inline. I'm also generating a new version of the paper:

> ## Typos & such
>
>   * p. 1, line 16: I think the word "for" needs to be inserted after "suitable".

Good catch. Fixed.

>   * p. 3, line 58: "celluar" → "cellular".

Fixed

> ## Suggestions (optional)

>   * In the section on related work ("What scribl is not"), I think it would enhance the paper to add a sentence or two contrasting scribl to Cytoscape. Although Cytoscape is indirectly referenced in the abstract by referencing the paper by Suderman & Hallett, I suspect that many readers working in this domain are likely to know about Cytoscape for graph visualization, and may wonder how scribl relates to it. IMHO, a direct explanation in the authors' own words would be preferable to making readers work it out for themselves.

Good point - I have added a sentence contextualizing the possible use of scribl Neo4j in conjunction with Cytoscape in the section about Reactome.  In the same section I've  also added a note that `scribl` could also be extended to generate models directly with enough annotations.

>   * Although the paper shows examples of scribl syntax used in adding tags to Zotero entries, it's not immediately obvious to readers how to make use of this data. The README for the repository does include examples of how to actually run scribl; still, in reading this paper, I kept thinking "okay, but what do I _do_ with this?" and had to go to the repository to find out. So I think another enhancement to the paper would be to add just one example of running scribl on the command line. It can be simply words to the effect of "After installing and configuring scribl, it can be simply invoked as a shell command, as shown in the following basic example", followed by a code block with a one-line scribl shell command. It would make the paper slightly more self-contained and help communicate to readers that usage is simple.

Agreed.  I have added a command-line sample in the section about the generation of the graph using `NetworkX`.

>   * In figure 2, the font size of the text inside the boxes of the figure is _very_ small, verging on microscopically small. Yes, readers can zoom in to read it, but it looks like there's enough room to enlarge the text a little bit. It doesn't have to be made huge, but IMHO a small enlargement would help. Figure 1 could use the same treatment; figure 3 is okay.

I have tried increasing the Fig 1 & 2 sizes by trimming the whitespace in the originals and increasing the width. In the case of Fig 2, I think I'm at about the limit where it gets too wide for the page (JOSS doesn't like to use the full page width - it seems to lose about 1/3 of the page due to the left margin). So it's at about the limit I can do without going back to the original figures and increasing the font size.  Hopefully it's sufficient now.
