### 2024-06-27 83008d7
slightly increase size of Fig 3
