+++ paper/paper.md
@@ -45,7 +45,7 @@ graph-based relationship queries.
 
 # The `scribl` language
 
-![The scribl schema comprises a hierarchy of five basic entities: `article`, `category`, `resource`, `process`, and `agent`. Here we depict an example network of entities and possible relationships for a single `article`.\label{fig:scribl-schema}](scribl-schema.png){ width=92% }
+![The scribl schema comprises a hierarchy of five basic entities: `article`, `category`, `resource`, `process`, and `agent`. Here we depict an example network of entities and possible relationships for a single `article`.\label{fig:scribl-schema}](scribl-schema.png){ width=100% }
 
 The language was designed for the curation of scientific articles to
 document the relationships between biological agents and processes
