### 2024-06-28 #6645 @alexlancaster
> > * In figure 2, the font size of the text inside the boxes of the figure is _very_ small, verging on microscopically small. Yes, readers can zoom in to read it, but it looks like there's enough room to enlarge the text a little bit. It doesn't have to be made huge, but IMHO a small enlargement would help. Figure 1 could use the same treatment; figure 3 is okay.

> I have tried increasing the Fig 1 & 2 sizes by trimming the whitespace in the originals and increasing the width. In the case of Fig 2, I think I'm at about the limit where it gets too wide for the page (JOSS doesn't like to use the full page width - it seems to lose about 1/3 of the page due to the left margin). So it's at about the limit I can do without going back to the original figures and increasing the font size. Hopefully it's sufficient now.

@mhucka we went back to the source and have replaced Figs 1 & 2 with versions with larger font size that should be clearer now. I'll use editorialbot to generate new versions of the paper in the next comment.
