### 2024-06-28 amberbiology/scribl#15 @benlansdell
Missing data for units tests/How to run unit tests
I'm unable to run the unit tests for this repo.

When running `pytest` I get many errors like
```
./tests/test_manage_graphdb.py::test_start Failed: [undefined]FileNotFoundError: [Errno 2] No such file or directory: 'tests/test_sandbox'
def test_start():
        print('\n\nTesting management of graph DB ...')
        # empty sandbox directory
>       empty_sandbox()

tests/test_manage_graphdb.py:28:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

    def empty_sandbox():
>       for tempfile in os.listdir(test_sandbox_dir):
E       FileNotFoundError: [Errno 2] No such file or directory: 'tests/test_sandbox'

tests/test_manage_graphdb.py:18: FileNotFoundError
```

Are there additional steps/files to download to get the unit tests working? I don't see this documented anywhere.

Or can these files be included in the repo if they are indeed missing?

Thanks!
