### 2024-06-28 amberbiology/scribl#15 @alexlancaster
Ah, yes.

`mkdir test_sandbox`

should be run before running `pytest`.  It's done in the CI runs.

I'll add that to the README.md.
