### 2024-06-28 #6645 @alexlancaster
> Hi @alexlancaster ,
>
> Apologies for the delay in getting to this. I've gone over the code and repo -- looks like a useful package! In going through the review checklist two items stand out as needing some attention. One I made an issue for in the main repo, the other I'll just mention here.
>
> The first is that I'm unable to run the unit tests locally myself. Many tests complain about files missing... (See issue https://github.com/amberbiology/scribl/issues/15)

Good catch, are my followup on the issue.

> The other is that I think the API documentation in the `scribl.pdf` (page 39 onwards) could benefit from some more standard formatting. The text description, I think, does consistently describe what the inputs and returns are, but I can't be certain. It would be useful to have standard argument lists/return types/etc for this.

Good point. I'll take a look at this. (Tagging my coauthor @gwebster on this as well).
