### 2024-06-28 f06ab59
add notes on running `pytest` (#15)
