### 2024-06-29 bc0527d
use `pytest`'s `tmpdir_factory` for sandbox
this avoids having to manually create a `test_sandbox` directory (#15)
