+++ tests/test_manage_graphdb.py
@@ -1,44 +1,40 @@
 __author__ = 'Amber Biology'
 
 
-import scribl, os, datetime, shutil, time
+import scribl
+import os
+import datetime
+import shutil
+import time
+import pytest
 from xml.etree import ElementTree as ET  # for GraphML checking
 from scribl.manage_graphdb import GraphDBInstance, generate_timestamp
 
 test_data_dir = os.path.join('tests', 'test_data')
-test_sandbox_dir = os.path.join('tests', 'test_sandbox')
-test_db_dir = os.path.join(test_sandbox_dir, 'test_graphdb')
 test_data_file = 'zotero_export_1.csv'
 updated_test_data_file = 'zotero_export_2.csv'
 zotero_csv_data = os.path.join(test_data_dir, test_data_file)
 updated_csv_data = os.path.join(test_data_dir, updated_test_data_file)
 
-# utility function to empty sandbox directory
-def empty_sandbox():
-    for tempfile in os.listdir(test_sandbox_dir):
-        rmfilepath = os.path.join(test_sandbox_dir, tempfile)
-        try:
-            os.remove(rmfilepath)
-        except:
-            shutil.rmtree(rmfilepath)
+# fixture function to create unique sandbox directories
+@pytest.fixture(scope="function")
+def sandbox_paths(tmpdir_factory):
+    test_sandbox_dir = tmpdir_factory.mktemp("scribl_sandbox")
+    test_db_dir = os.path.join(test_sandbox_dir, 'test_graphdb')
+    return test_sandbox_dir, test_db_dir
 
-def test_start():
-    print('\n\nTesting management of graph DB ...')
-    # empty sandbox directory
-    empty_sandbox()
-
-def test_graphdb_initialization():
+def test_graphdb_initialization(sandbox_paths):
     print('Testing graph DB initialization ...')
-    empty_sandbox()
+    test_sandbox_dir, test_db_dir = sandbox_paths
     gdb = GraphDBInstance(test_db_dir)
     assert os.path.exists(test_db_dir)
     for folder in ['backup', 'config', 'db_snapshots', 'zotero_csv_exports']:
         folder_path = os.path.join(test_db_dir, folder)
         assert os.path.exists(folder_path)
 
-def test_graphdb_metadata():
+def test_graphdb_metadata(sandbox_paths):
     print('Testing graph DB metadata ...')
-    empty_sandbox()
+    test_sandbox_dir, test_db_dir = sandbox_paths
     gdb = GraphDBInstance(test_db_dir)
     db_name = 'Systems Biology Literature DB for testing'
     curator = 'Amber Biology (info@amberbiology.com)'
@@ -66,9 +62,9 @@ def test_graphdb_metadata():
     assert lines[2][36:].strip() == "That's what she said"
     assert len(gdb.generate_metadata_cypher()) == 653
 
-def test_graphdb_import():
+def test_graphdb_import(sandbox_paths):
     print('Testing graph DB import ...')
-    empty_sandbox()
+    test_sandbox_dir, test_db_dir = sandbox_paths
     gdb = GraphDBInstance(test_db_dir)
     db_name = 'Systems Biology Literature DB for testing'
     curator = 'Amber Biology (info@amberbiology.com)'
@@ -91,9 +87,9 @@ def test_graphdb_import():
     assert len(gdb.graphdb.db['article']) == 13
 
 
-def test_graphdb_reimport():
+def test_graphdb_reimport(sandbox_paths):
     print('Testing graph DB re-import ...')
-    empty_sandbox()
[…]
