### 2024-06-29 7fa9029
hardcode the full beta version
