+++ .github/workflows/python-package.yml
@@ -123,7 +123,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.13", "3.12", "3.11", "3.10", "3.9"]
+        python-version: ["3.13.0-beta.2", "3.12", "3.11", "3.10", "3.9"]
 
     steps:
     - uses: actions/checkout@v4
