### 2024-06-29 d2e9bb9
better: supply the "allow-prerelease" in action
