### 2024-07-01 #6645 @mhucka
> @mhucka we went back to the source and have replaced Figs 1 & 2 with versions with larger font size that should be clearer now. I'll use editorialbot to generate new versions of the paper in the next comment.

The new figures and font sizes are great! Thanks for doing that.
