### 2024-07-02 c0d55ae
use `pytest`'s `tmpdir_factory` for sandbox (#16)
this avoids having to manually create a `test_sandbox` directory (#15), thanks to @benlansdell for the report
