### 2024-07-02 bdb0fdc
updated API documentation to include more details
addressing: https://github.com/openjournals/joss-reviews/issues/6645#issuecomment-2200599628
