### 2024-07-03 #6645 @benlansdell
> > @alexlancaster Thanks for the quick response. I can confirm the PR you made fixes the testing issue. If the documentation is fixed up slightly, I'm happy with everything else
>
> Great - @gwebster and I worked on the method docs, and the PDF is now updated.

Looks good! Thanks for fixing that up quickly. That is all of my concerns addressed.
