### 2024-07-08 e8cae43
documentation updates
* affiliation for Zenodo

* full link to documentation file

* add pypi target
