+++ CITATION.cff
@@ -13,10 +13,9 @@ authors:
   - family-names: Lancaster
     given-names: Alexander K.
     orcid: https://orcid.org/0000-0002-0002-9263
-    affiliation: Amber Biology LLC and Ronin Institute
+    affiliation: Amber Biology LLC and Institute for Globally Distributed Open Research and Education (IGDORE)
 title: >-
-  scribl: A system for the semantic capture of relationships in biological
-  literature
+  scribl: A system for the semantic capture of relationships in biological literature
 url: https://amberbiology.com
 repository-code: https://github.com/amberbiology/scribl
 type: software
