+++ .github/workflows/python-package.yml
@@ -196,7 +196,7 @@ jobs:
     environment: pypi
     permissions:
       id-token: write
-    if: github.event.release.target_commitish == 'main' && github.event_name == 'release' && github.event.action == 'published' && always()  && !cancelled() && !failure()
+    if: github.event.release.target_commitish == 'pypi_zenodo_sandbox' && github.event_name == 'release' && github.event.action == 'published' && always()  && !cancelled() && !failure()
     steps:
       - uses: actions/download-artifact@v4
         with:
