+++ .github/workflows/python-package.yml
@@ -58,7 +58,7 @@ jobs:
     runs-on: ubuntu-latest
     # "else" we *are* publishing something, only works for non-pre-releases on the main branch
     # if: false
-    if: github.event.release.target_commitish == 'main' && github.event_name == 'release' && github.event.action == 'published' && github.event.release.prerelease == false
+    if: github.event.release.target_commitish == 'pypi_zenodo_sandbox' && github.event_name == 'release' && github.event.action == 'published' && github.event.release.prerelease == false
     steps:
       - name: Checkout the contents of your repository
         uses: actions/checkout@v4
