### 2024-07-08 36a2023
Upserted version into CITATION.cff and into .zenodo.json
