+++ CITATION.cff
@@ -1,6 +1,10 @@
 cff-version: 1.2.0
 abstract: >-
-  A system for the semantic capture of relationships in biological literature. The **scribl** language was designed for the curation from scientific articles, of the relationships between the various biological agents and processes that they describe, with a view to generating a graph database that captures these relationships as a connected network.
+  A system for the semantic capture of relationships in biological literature.
+  The **scribl** language was designed for the curation from scientific
+  articles, of the relationships between the various biological agents and
+  processes that they describe, with a view to generating a graph database that
+  captures these relationships as a connected network.
 authors:
   - family-names: Webster
     given-names: Gordon
@@ -9,9 +13,12 @@ authors:
   - family-names: Lancaster
     given-names: Alexander K.
     orcid: https://orcid.org/0000-0002-0002-9263
-    affiliation: Amber Biology LLC and Institute for Globally Distributed Open Research and Education (IGDORE)
+    affiliation: >-
+      Amber Biology LLC and Institute for Globally Distributed Open Research and
+      Education (IGDORE)
 title: >-
-  scribl: A system for the semantic capture of relationships in biological literature
+  scribl: A system for the semantic capture of relationships in biological
+  literature
 url: https://amberbiology.com
 repository-code: https://github.com/amberbiology/scribl
 type: software
@@ -23,3 +30,4 @@ keywords:
   - computational biology
 license: AGPL-3.0-or-later
 version: v0.7.16
+doi: 10.5281/zenodo.80499
