### 2024-07-08 d6df6dc
don't use `--find-links` in testing
attempts to pull in version from PyPI
