### 2024-07-08 a8dc424
force use of bash in step for pip install
