### 2024-07-08 46a1bba
Fix `pip install` during `pytest` (#17)
* force use of bash on all platforms, when running `pytest`, so wildcard expansion works on Windows (the default powershell on Windows, fails)
