### 2024-07-08 6c2fcf5
Upserted version into CITATION.cff and into .zenodo.json
