+++ CITATION.cff
@@ -1,14 +1,6 @@
 cff-version: 1.2.0
 abstract: >-
-  scribl (Semantic Capture of Relationships in Biological Literature)
-  is a Python software pipeline and API that transforms a Zotero
-  literature database with entries annotated with tags in the
-  **scribl** language to create a graph data structure, suitable for
-  import into a graph database platform, such as Neo4j or Python's
-  NetworkX. The scribl language was designed for the curation from
-  scientific articles, to capture the relationships between the
-  various biological agents and the processes that they describe in a
-  connected network.
+  scribl (Semantic Capture of Relationships in Biological Literature) is a Python software pipeline and API that transforms a Zotero literature database with entries annotated with tags in the **scribl** language to create a graph data structure, suitable for import into a graph database platform, such as Neo4j or Python's NetworkX. The scribl language was designed for the curation from scientific articles, to capture the relationships between the various biological agents and the processes that they describe in a connected network.
 authors:
   - family-names: Webster
     given-names: Gordon
@@ -18,11 +10,9 @@ authors:
     given-names: Alexander K.
     orcid: https://orcid.org/0000-0002-0002-9263
     affiliation: >-
-      Amber Biology LLC and Institute for Globally Distributed Open Research and
-      Education (IGDORE)
+      Amber Biology LLC and Institute for Globally Distributed Open Research and Education (IGDORE)
 title: >-
-  scribl: A system for the semantic capture of relationships in biological
-  literature
+  scribl: A system for the semantic capture of relationships in biological literature
 url: https://amberbiology.com
 repository-code: https://github.com/amberbiology/scribl
 type: software
@@ -33,5 +23,5 @@ keywords:
   - bioinformatics
   - computational biology
 license: AGPL-3.0-or-later
-version: v0.7.16
+version: v0.7.17
 doi: 10.5281/zenodo.80499
