### 2024-07-08 b44f461
Update pyzotero requirement from <=1.5.19 to <=1.5.20
Updates the requirements on [pyzotero](https://github.com/urschrei/pyzotero) to permit the latest version.
- [Release notes](https://github.com/urschrei/pyzotero/releases)
- [Commits](https://github.com/urschrei/pyzotero/compare/v0.3...v1.5.20)

---
updated-dependencies:
- dependency-name: pyzotero
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
