### 2024-07-08 996e873
Update matplotlib requirement from <=3.9.0 to <=3.9.1
Updates the requirements on [matplotlib](https://github.com/matplotlib/matplotlib) to permit the latest version.
- [Release notes](https://github.com/matplotlib/matplotlib/releases)
- [Commits](https://github.com/matplotlib/matplotlib/compare/v0.91.2...v3.9.1)

---
updated-dependencies:
- dependency-name: matplotlib
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
