+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# Changelog
+
+All notable changes to this project (especially user-visible ones)
+will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org).
