### 2024-07-11 86f1b00
update CHANGELOG.md
link to CHANGELOG.md
