+++ CHANGELOG.md
@@ -5,3 +5,41 @@ will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org).
+
+## [0.8.0] - 2024-07-11
+
+First production release of `scribl`
+
+### Features
+
+- Add support for zenodo and `pypi` uploads (disabled for the time being) by @alexlancaster ([#3](https://github.com/amberbiology/scribl/pull/3))
+
+### Bug Fixes
+
+- Fix `pip install` during `pytest` by @alexlancaster ([#17](https://github.com/amberbiology/scribl/pull/17))
+- create sandbox for `pytest` "on-the-fly", rather than hardcoding by @alexlancaster ([#16](https://github.com/amberbiology/scribl/pull/16))
+- implement best-practice conventions as per `repo-review` by @alexlancaster ([#2](https://github.com/amberbiology/scribl/pull/2))
+
+### Internal
+
+- Update matplotlib requirement from \<=3.9.0 to \<=3.9.1 by @dependabot ([#19](https://github.com/amberbiology/scribl/pull/19))
+- Update pyzotero requirement from \<=1.5.19 to \<=1.5.20 by @dependabot ([#18](https://github.com/amberbiology/scribl/pull/18))
+- Update matplotlib requirement from \<=3.8.4 to \<=3.9.0 by @dependabot ([#13](https://github.com/amberbiology/scribl/pull/13))
+- Update pyzotero requirement from \<=1.5.18 to \<=1.5.19 by @dependabot ([#14](https://github.com/amberbiology/scribl/pull/14))
+- Update pandas requirement from \<=2.2.1 to \<=2.2.2 by @dependabot ([#12](https://github.com/amberbiology/scribl/pull/12))
+- Update networkx requirement from \<=3.2.1 to \<=3.3 by @dependabot ([#11](https://github.com/amberbiology/scribl/pull/11))
+- Update matplotlib requirement from \<=3.8.3 to \<=3.8.4 by @dependabot ([#10](https://github.com/amberbiology/scribl/pull/10))
+- Update pyparsing requirement from \<=3.1.1 to \<=3.1.2 by @dependabot ([#9](https://github.com/amberbiology/scribl/pull/9))
+- Update Zenodo + Test PyPI workflows by @alexlancaster ([#8](https://github.com/amberbiology/scribl/pull/8))
+- Update pandas requirement from \<=2.2.0 to \<=2.2.1 by @dependabot ([#7](https://github.com/amberbiology/scribl/pull/7))
+- Update matplotlib requirement from \<=3.8.2 to \<=3.8.3 by @dependabot ([#6](https://github.com/amberbiology/scribl/pull/6))
+- Bump the actions group with 1 update by @dependabot ([#5](https://github.com/amberbiology/scribl/pull/5))
+- Update pandas requirement from \<=2.1.4 to \<=2.2.0 by @dependabot ([#4](https://github.com/amberbiology/scribl/pull/4))
+- Add support for zenodo and `pypi` uploads (disabled for the time being) by @alexlancaster ([#3](https://github.com/amberbiology/scribl/pull/3))
+- Bump actions/upload-artifact from 3 to 4 by @dependabot ([#1](https://github.com/amberbiology/scribl/pull/1))
+
+### Documentation
+
+- implement best-practice conventions as per `repo-review` by @alexlancaster ([#2](https://github.com/amberbiology/scribl/pull/2))
+
+[0.8.0]: https://github.com/amberbiology/scribl/releases/tag/v0.8.0
