### 2024-07-11 aff3d1e
Upserted version into CITATION.cff and into .zenodo.json
