### 2024-07-11 #6645 @alexlancaster
Thank you @atrisovic !

> ## Additional Author Tasks After Review is Complete
>
>   * Double check authors and affiliations (including ORCIDs)

All correct

>   * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Release made: v0.8.0:

https://github.com/amberbiology/scribl/releases/tag/v0.8.0

>   * Archive the release on Zenodo/figshare/etc and post the DOI here.

v0.8.0 archived on Zenodo:

https://zenodo.org/records/12728363  / DOI: 10.5281/zenodo.12728363

The DOI for the "all versions": 10.5281/zenodo.12728362

>   * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

They match.

>   * Make sure that the license listed for the archive is the same as the software license.

Matched: "GNU Affero General Public License v3.0 or later" (AGPLv3.0+)
