### 2024-07-11 #6645 @atrisovic
Hi @alexlancaster I have now created a pull request with a few edits https://github.com/amberbiology/scribl/pull/21.
