### 2024-07-11 #6645 @alexlancaster
> Hi @alexlancaster I have now created a pull request with a few edits [amberbiology/scribl#21](https://github.com/amberbiology/scribl/pull/21).

Thanks, now merged into my `manuscript` branch. I also went through the checklist above.
