### 2024-07-11 9649bca
update README with publication details
link to changelog, note PyPI package is now available
