### 2024-07-12 #6645 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
