### 2024-07-12 #6645 @atrisovic
Nothing for the authors, the paper now needs final approval from @openjournals/bcm-eics
Congrats! :)
