### 2024-07-17 #6645 @Kevin-Mattheus-Moerman
Congratulations @alexlancaster et al. for this JOSS publication!

Thanks for editing @atrisovic !
